Change the label selection based on current behavior
The releases created and their labels seem to be different today than
in the past.  I no longer see pull request releases and now the
releases that are all push based have some different event-types like
incoming for rebuild requests and retest-comment for retest requests.
I'm not sure I want to focus on that label anymore so switching for
the momment to a different label, but this is something to keep an eye
on.

Signed-off-by: Gus Parvin <gparvin@redhat.com>